component-entrypoints
- `arrow-button`: добавлена точка входа `index.ts` (patch)
- `button`: добавлена точка входа `index.ts` (patch)
- `checkbox`: добавлена точка входа `index.ts` (patch)
- `dot-nav`: добавлена точка входа `index.ts` (patch)
- `expandable`: добавлена точка входа `index.ts` (patch)
- `field-block`: добавлена точка входа `index.ts` (patch)
- `types`: интерфейс `CheckboxInputProps` теперь не содержит поля `type` (patch)
