+++ src/arrow-button/index.ts
@@ -1,0 +1,2 @@
+export type { ArrowButtonSize, ArrowDirection, ArrowButtonProps } from './types';
+export { ArrowButton } from './arrow-button';
