+++ src/checkbox/index.ts
@@ -1,0 +1,2 @@
+export type { CheckboxProps } from './types';
+export { Checkbox } from './checkbox';
