+++ src/dot-nav/index.ts
@@ -1,0 +1,2 @@
+export type { DotNavSize, DotNavProps } from './types';
+export { DotNav } from './dot-nav';
