+++ src/dot-nav/types.ts
@@ -1,0 +1,15 @@
+export type DotNavSize = 's' | 'l';
+
+export interface DotNavProps {
+  /** Индекс выбранной точки. */
+  current?: number;
+
+  /** Количество точек. */
+  total?: number;
+
+  /** Сработает при выборе точки. */
+  onSelect?: (index: number) => void;
+
+  /** Размер. */
+  size?: DotNavSize;
+}
