+++ src/expandable/index.ts
@@ -1,0 +1,2 @@
+export type { ExpandableGroupProps, ExpandableGroupItemProps } from './types';
+export { Expandable, ExpandableGroup, ExpandableItem } from './expandable-group';
