+++ src/field-block/index.ts
@@ -1,0 +1,2 @@
+export type { FieldBlockSize, FieldBlockStyle, FieldBlockProps } from './types';
+export { FieldBlock, FIELD_BLOCK_DEFAULTS, FIELD_BLOCK_HEIGHT } from './field-block';
